requirements.txt
ns1-python
requests
